+++ fabric/src/main/java/de/maxhenkel/voicechat/mixin/PlayerEntryMixin.java
@@ -59,7 +59,7 @@ public class PlayerEntryMixin {
     @Redirect(method = "<init>", at = @At(value = "INVOKE", target = "Lcom/google/common/collect/ImmutableList;of(Ljava/lang/Object;Ljava/lang/Object;Ljava/lang/Object;)Lcom/google/common/collect/ImmutableList;"))
     private ImmutableList<?> children(Object o1, Object o2, Object o3) {
         inviteButton = new ImageButton(0, 0, GROUP_ICON, button -> {
-            minecraft.player.commandUnsigned("voicechat invite %s".formatted(playerName));
+            minecraft.player.connection.sendUnsignedCommand("voicechat invite %s".formatted(playerName));
             invited = true;
         }, (button, matrices, mouseX, mouseY) -> {
             if (screen == null || invited) {