+++ fabric/changelog.md
@@ -1,1 +1,2 @@
-- Fade players volume when dying
+- Updated to 22w42a
+- Made voice chat resource pack names translatable